requirements.txt
Flask
gunicorn
openai
langchain
python-dotenv
requests
google-search-results
streamlit
streamlit_chat
tqdm>=4.65.0
black>=23.3.0
flake8>=6.0.0
isort>=5.12.0
pytest>=7.3.2
tiktoken>=0.4.0
chromadb
faiss-cpu
sympy
pillow